[#71]/refactor ProfileNicknameForm, ProfileEditButtonGroup 컴포넌트 추출
